convert int to float writable

+++ src/com/agarsofttech/mr/GACombiner.java
@@ -2,19 +2,20 @@ package com.agarsofttech.mr;
 
 import java.io.IOException;
 
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
 public class GACombiner
-       extends Reducer<Text,IntWritable,Text,IntWritable> {
-    private IntWritable result = new IntWritable();
+       extends Reducer<Text,FloatWritable,Text,FloatWritable> {
+    private FloatWritable result = new FloatWritable();
 
-    public void reduce(Text key, Iterable<IntWritable> values,
+    public void reduce(Text key, Iterable<FloatWritable> values,
                        Context context
                        ) throws IOException, InterruptedException {
       int sum = 0;
-      for (IntWritable val : values) {
+      for (FloatWritable val : values) {
         sum += val.get();
       }
       /*if(isRuleFit(sum))*/
