+++ src/com/agarsofttech/mr/SPGATry1.java
@@ -2,7 +2,8 @@ package com.agarsofttech.mr;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.FloatWritable;
+//import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -18,7 +19,7 @@ public static void main(String[] args) throws Exception {
     job.setCombinerClass(GACombiner.class);
     job.setReducerClass(GAReducer.class);
     job.setOutputKeyClass(Text.class);
-    job.setOutputValueClass(IntWritable.class);
+    job.setOutputValueClass(FloatWritable.class);
     //FileInputFormat.addInputPath(job, new Path(args[0]));
     //FileOutputFormat.setOutputPath(job, new Path(args[1]));
     FileInputFormat.addInputPath(job, new Path("data/input1.txt"));
